+++ Code/Gui/MITK/Modules/niftkOpenCV/VideoTrackerMatching/mitkVideoTrackerMatching.h
@@ -57,7 +57,7 @@ public:
   /**
    * \brief Return the tracking matrix for a given video frame number
    */
-  cv::Mat GetTrackerMatrix ( unsigned int FrameNumber, int * TimingError = NULL, unsigned int TrackerIndex = 0 );
+  cv::Mat GetTrackerMatrix ( unsigned int FrameNumber, long * TimingError = NULL, unsigned int TrackerIndex = 0 );
 
   /**
    * \brief returns state of m_Ready
